Breaking transformer modules.py into smaller files (#3466)
* Separating out components of transformer into their own files

* move the rest of transformer modules out of modules.py

* remove transformer modules.py

* resolving imports

* move functions into modules, fix imports

* make a constants.py

* noqa comment placement 🙄
