+++ parlai/agents/transformer/modules/__init__.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .constants import LAYER_NORM_EPS  # noqa: F401
+from .functions import (  # noqa: F401
+    create_embeddings,
+    create_position_codes,
+    get_n_positions_from_options,
+)
+from .attention import BasicAttention, MultiHeadAttention  # noqa: F401
+from .ffn import TransformerFFN  # noqa: F401
+from .encoder import TransformerEncoder, TransformerEncoderLayer  # noqa: F401
+from .decoder import TransformerDecoder, TransformerDecoderLayer  # noqa: F401
+from .generator import TransformerGeneratorModel  # noqa: F401
+from .wrappers import TransformerLinearWrapper, TransformerResponseWrapper  # noqa: F401
+from .mem_net import TransformerMemNetModel  # noqa: F401
