+++ parlai/agents/transformer/modules/attention.py
@@ -1,0 +1,251 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+Implementations of attention.
+"""
+
+import math
+from typing import Dict, Tuple, Optional
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+from parlai.utils.torch import neginf
+
+
+class BasicAttention(nn.Module):
+    """
+    Implements simple/classical attention.
+    """
+
+    def __init__(self, dim=1, attn='cosine', residual=False, get_weights=True):
+        super().__init__()
+        if attn == 'cosine':
+            self.cosine = nn.CosineSimilarity(dim=dim)
+        self.attn = attn
+        self.dim = dim
+        self.get_weights = get_weights
+        self.residual = residual
+
+    def forward(self, xs, ys, mask_ys=None, values=None):
+        """
+        Compute attention.
+
+        Attend over ys with query xs to obtain weights, then apply weights to
+        values (ys if yalues is None)
+
+        Args:
+            xs: B x query_len x dim (queries)
+            ys: B x key_len x dim (keys)
+            mask_ys: B x key_len (mask)
+            values: B x value_len x dim (values); if None, default to ys
+        """
+        bsz = xs.size(0)
+        y_len = ys.size(1)
+        x_len = xs.size(1)
+        if self.attn == 'cosine':
+            l1 = self.cosine(xs, ys).unsqueeze(self.dim - 1)
+        else:
+            l1 = torch.bmm(xs, ys.transpose(1, 2))
+            if self.attn == 'sqrt':
+                d_k = ys.size(-1)
+                l1 = l1 / math.sqrt(d_k)
+        if mask_ys is not None:
+            attn_mask = (mask_ys == 0).view(bsz, 1, y_len)
+            attn_mask = attn_mask.repeat(1, x_len, 1)
+            l1.masked_fill_(attn_mask, neginf(l1.dtype))
+        l2 = F.softmax(l1, dim=self.dim, dtype=torch.float).type_as(l1)
+        if values is None:
+            values = ys
+        lhs_emb = torch.bmm(l2, values)
+
+        # # add back the query
+        if self.residual:
+            lhs_emb = lhs_emb.add(xs)
+
+        if self.get_weights:
+            return lhs_emb.squeeze(self.dim - 1), l2
+        else:
+            return lhs_emb.squeeze(self.dim - 1)
+
+
+class MultiHeadAttention(nn.Module):
+    """
+    Implements MultiHeadAttention; this is the core workhorse of the Transformer.
+
+    See Vaswani (2017) for an extensive description.
+    """
+
+    def __init__(self, n_heads, dim, dropout=0):
+        super(MultiHeadAttention, self).__init__()
+        self.n_heads = n_heads
+        self.dim = dim
+
+        self.attn_dropout = nn.Dropout(p=dropout)  # --attention-dropout
+        self.q_lin = nn.Linear(dim, dim)
+        self.k_lin = nn.Linear(dim, dim)
+        self.v_lin = nn.Linear(dim, dim)
+        # TODO: merge for the initialization step
+        nn.init.xavier_normal_(self.q_lin.weight)
+        nn.init.xavier_normal_(self.k_lin.weight)
+        nn.init.xavier_normal_(self.v_lin.weight)
+        # and set biases to 0
+        self.out_lin = nn.Linear(dim, dim)
+
+        nn.init.xavier_normal_(self.out_lin.weight)
+
+    def forward(  # type: ignore
+        # TODO: remove type ignore with pytorch 1.5:
+        # https://github.com/pytorch/pytorch/pull/31057
+        self,
+        query: torch.Tensor,
+        key: Optional[torch.Tensor] = None,
+        value: Optional[torch.Tensor] = None,
+        mask: torch.Tensor = None,
+        incr_state: Optional[Dict[str, torch.Tensor]] = None,
+        static_kv: bool = False,
+    ) -> Tuple[torch.Tensor, Dict[str, torch.Tensor], torch.Tensor]:
+        """
+        Forward pass.
+
+        :param query: attention query
+        :param key: attention key
+        :param value: attention value
+        :param mask: tensor in which True means that we are allowing attention and False
+          means we are blocking it. Mask is:
+          - [B, key_len] (encoder self-attn and decoder enc/dec attn)
+          - [B, query_len, key_len] (decoder self-attn)
+          - [B, 1, key_len] (decoder self-attn with incr_state caching)
+        :param incr_state: dictionary with values representing the previous states of
+          the key, value, and mask
+        :param static_kv: True if the key and value are held constant during decoding
+          (as in encoder/decoder attention)
+        :return: (
+          final attended tensor,
+          new incremental state,
+          key/value-multiplied tensor before softmax,
+        )
+        """
+
+        batch_size, query_len, dim = query.size()
+        assert (
+            dim == self.dim
+        ), 'Dimensions do not match: {} query vs {} configured'.format(dim, self.dim)
+        assert mask is not None, 'Mask is None, please specify a mask'
+        n_heads = self.n_heads
+        dim_per_head = dim // n_heads
+        scale = math.sqrt(dim_per_head)
+
+        def prepare_head(tensor):
+            # input is [batch_size, seq_len, n_heads * dim_per_head]
+            # output is [batch_size * n_heads, seq_len, dim_per_head]
+            bsz, seq_len, _ = tensor.size()
+            tensor = tensor.view(batch_size, tensor.size(1), n_heads, dim_per_head)
+            tensor = (
+                tensor.transpose(1, 2)
+                .contiguous()
+                .view(batch_size * n_heads, seq_len, dim_per_head)
+            )
+            return tensor
+
+        # q, k, v are the transformed values
+        if key is None and value is None:
+            # self attention
+            key = value = query
+            _, _key_len, dim = query.size()
+        elif value is None:
+            # key and value are the same, but query differs
+            # self attention
+            value = key
+
+        assert key is not None  # let mypy know we sorted this
+        _, _key_len, dim = key.size()
+
+        q = prepare_head(self.q_lin(query))
+        k = prepare_head(self.k_lin(key))
+        v = prepare_head(self.v_lin(value))
+
+        # Prepend incremental states. For each of the key, value, and mask, see if
+        # a previous incremental state exists, and if so, reshape it to match the shape
+        # of the new state. Concatenate the previous and new states to match what the
+        # full state would have been if we had not cached. (If we are using static_kv,
+        # these three states are unchanging, so just re-use the cached states.)
+        if incr_state is None:
+            incr_state = {}
+        if 'prev_key' in incr_state:
+            prev_key = incr_state['prev_key'].view(
+                batch_size * n_heads, -1, dim_per_head
+            )
+            if static_kv:
+                k = prev_key
+            else:
+                k = torch.cat([prev_key, k], dim=1)
+        if 'prev_value' in incr_state:
+            prev_value = incr_state['prev_value'].view(
+                batch_size * n_heads, -1, dim_per_head
+            )
+            if static_kv:
+                v = prev_value
+            else:
+                v = torch.cat([prev_value, v], dim=1)
+        if 'prev_mask' in incr_state:
+            if static_kv:
+                mask = incr_state['prev_mask']
+            else:
+                mask = torch.cat([incr_state['prev_mask'], mask], dim=2)
+                # Prepend along the key_len dimension (analogous to
+                # incr_state['prev_key'])
+
+        # Save new incremental states. We reshape to allow for reordering along batch
+        # dimension.
+        new_incr_state = {
+            'prev_key': k.view(batch_size, n_heads, -1, dim_per_head),
+            'prev_value': v.view(batch_size, n_heads, -1, dim_per_head),
+            'prev_mask': mask,
+        }
+
+        full_key_len = k.size(1)
+        dot_prod = q.div_(scale).bmm(k.transpose(1, 2))
+        # [B * n_heads, query_len, key_len]
+        attn_mask = (
+            (mask == 0)
+            .view(batch_size, 1, -1, full_key_len)
+            .repeat(1, n_heads, 1, 1)
+            .expand(batch_size, n_heads, query_len, full_key_len)
+            .view(batch_size * n_heads, query_len, full_key_len)
+        )
+        assert attn_mask.shape == dot_prod.shape
+        dot_prod.masked_fill_(attn_mask, neginf(dot_prod.dtype))
+
+        attn_weights = F.softmax(
+            dot_prod, dim=-1, dtype=torch.float  # type: ignore
+        ).type_as(query)
+        attn_weights = self.attn_dropout(attn_weights)  # --attention-dropout
+
+        attentioned = attn_weights.bmm(v)
+        attentioned = (
+            attentioned.type_as(query)
+            .view(batch_size, n_heads, query_len, dim_per_head)
+            .transpose(1, 2)
+            .contiguous()
+            .view(batch_size, query_len, dim)
+        )
+
+        out = self.out_lin(attentioned)
+
+        return out, new_incr_state, dot_prod
+
+    def reorder_incremental_state(
+        self, incremental_state: Dict[str, torch.Tensor], inds: torch.Tensor
+    ) -> Dict[str, torch.Tensor]:
+        """
+        Reorder the input incremental-state tensors.
+        """
+        return {
+            key: torch.index_select(val, 0, inds.to(val.device)).contiguous()
+            for key, val in incremental_state.items()
+        }
