+++ parlai/agents/transformer/modules/constants.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+LAYER_NORM_EPS = 1e-5  # Epsilon for layer norm.
