+++ parlai/agents/transformer/modules/encoder.py
@@ -1,0 +1,363 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+Transformer encoder implementations.
+"""
+
+from typing import Tuple, Optional, Union
+
+import numpy as np
+import torch
+import torch.nn as nn
+
+from parlai.agents.transformer.modules import (
+    create_position_codes,
+    get_n_positions_from_options,
+    LAYER_NORM_EPS,
+    MultiHeadAttention,
+    TransformerFFN,
+)
+from parlai.core.opt import Opt
+from parlai.utils.misc import warn_once
+from parlai.utils.torch import PipelineHelper
+
+
+class TransformerEncoder(nn.Module):
+    """
+    Transformer encoder module.
+
+    For documentation on parameters that are take directly from opt,
+    see parlai/agents/transformer/transformer.py
+
+    :param opt: ParlAI-parsed options.
+    :param vocabulary_size: Count of tokens/words in the dictionary.
+    :param embedding: an embedding matrix for the bottom layer of the transformer.
+        If none, one is created for this encoder.
+    :param int padding_idx: Reserved padding index in the embeddings matrix.
+    :param str reduction_type: Type of reduction at the end of the encoder.
+    :param int n_positions: Size of the position embeddings matrix.
+    :param int n_segments: Number of segments/lang/sentence embeddings.
+    :param bool embeddings_scale: Scale embeddings relative to their dimensionality.
+        Found useful in fairseq.
+    """
+
+    def __init__(
+        self,
+        opt: Opt,
+        vocabulary_size: int,
+        embedding: Optional[nn.Embedding] = None,
+        padding_idx: int = 0,
+        reduction_type: str = 'mean',
+        n_positions: Optional[int] = None,
+        n_segments: Optional[int] = None,
+        embeddings_scale: Optional[bool] = None,
+        dropout: Optional[float] = None,
+        activation: Optional[str] = None,
+        variant: Optional[str] = None,
+        output_scaling: Optional[float] = None,
+    ):
+        super(TransformerEncoder, self).__init__()
+
+        def _default(val, default):
+            return val if val is not None else default
+
+        self.embedding_size = opt['embedding_size']
+        self.ffn_size = opt['ffn_size']
+        self.n_layers = (
+            opt['n_encoder_layers']
+            if opt.get('n_encoder_layers', -1) > 0
+            else opt['n_layers']
+        )
+        self.n_heads = opt['n_heads']
+        self.dim = self.embedding_size
+        self.embeddings_scale = _default(
+            embeddings_scale, opt.get('embeddings_scale', False)
+        )
+        self.reduction_type = reduction_type
+        self.padding_idx = padding_idx
+        # this is --dropout, not --relu-dropout or --attention-dropout
+        self.dropout_frac = _default(dropout, opt.get('dropout', 0.0))
+        self.dropout = nn.Dropout(p=self.dropout_frac)
+        self.variant = _default(variant, opt.get('variant', 'aiayn'))
+        self.n_segments = _default(n_segments, opt.get('n_segments', 0))
+
+        self.n_positions = _default(n_positions, get_n_positions_from_options(opt))
+        self.out_dim = self.embedding_size
+        assert (
+            self.embedding_size % self.n_heads == 0
+        ), 'Transformer embedding size must be a multiple of n_heads'
+
+        # check input formats:
+        if embedding is not None:
+            assert (
+                self.embedding_size is None
+                or self.embedding_size == embedding.weight.shape[1]
+            ), "Embedding dim must match the embedding size."
+
+        if embedding is not None:
+            self.embeddings = embedding
+        else:
+            raise AssertionError(
+                "This code should not execute. Left here in case we want to enable it."
+            )
+            assert self.padding_idx is not None
+            self.embeddings = nn.Embedding(
+                vocabulary_size, self.embedding_size, padding_idx=padding_idx
+            )
+            nn.init.normal_(self.embeddings.weight, 0, self.embedding_size ** -0.5)
+
+        # create the positional embeddings
+        self.position_embeddings = nn.Embedding(self.n_positions, self.embedding_size)
+        if not opt.get('learn_positional_embeddings', False):
+            create_position_codes(
+                self.n_positions,
+                self.embedding_size,
+                out=self.position_embeddings.weight,
+            )
+        else:
+            nn.init.normal_(
+                self.position_embeddings.weight, 0, self.embedding_size ** -0.5
+            )
+
+        # embedding normalization
+        if (
+            self.variant == 'xlm'
+            or self.variant == 'prelayernorm'
+            or self.variant == 'bart'
+        ):
+            self.norm_embeddings = torch.nn.LayerNorm(self.dim, eps=LAYER_NORM_EPS)
+        elif self.variant == 'aiayn':
+            pass
+        else:
+            raise ValueError("Can't handle --variant {}".format(self.variant))
+
+        if self.n_segments >= 1:
+            self.segment_embeddings = nn.Embedding(self.n_segments, self.dim)
+
+        # build the model
+        self.layers = nn.ModuleList()
+        for _ in range(self.n_layers):
+            self.layers.append(
+                TransformerEncoderLayer(
+                    self.n_heads,
+                    self.embedding_size,
+                    self.ffn_size,
+                    attention_dropout=opt.get('attention_dropout', 0.0),
+                    relu_dropout=opt.get('relu_dropout', 0.0),
+                    dropout=self.dropout_frac,
+                    variant=self.variant,
+                    activation=_default(activation, opt.get('activation', 'relu')),
+                )
+            )
+        self.output_scaling = _default(output_scaling, opt.get('output_scaling', 1.0))
+
+    def forward_embedding(
+        self,
+        input: torch.LongTensor,
+        positions: Optional[torch.LongTensor] = None,
+        segments: Optional[torch.LongTensor] = None,
+    ) -> Tuple[torch.Tensor, torch.BoolTensor]:
+        """
+        Embed tokens prior to feeding into transformer.
+
+        :param LongTensor[batch,seqlen] input:
+            The input IDs
+        :param LongTensor[batch,seqlen] positions:
+            Positions for input IDs
+        :param LongTensor[batch,seqlen]:
+            If provided, additionally adds ``segments`` as extra embedding features.
+
+        :return (tensor, mask):
+            return embedded input and mask
+        """
+        mask = input != self.padding_idx
+        if positions is None:
+            positions = (mask.cumsum(dim=1, dtype=torch.int64) - 1).clamp_(min=0)
+        tensor = self.embeddings(input)
+        if self.embeddings_scale:
+            tensor = tensor * np.sqrt(self.dim)
+
+        if positions.max().item() > self.n_positions:
+            warn_once(
+                'You are inputting a sequence of {x} length, but only have '
+                '--n-positions {y}. Set --truncate or increase --n-positions'.format(
+                    x=positions.max().item(), y=self.n_positions
+                )
+            )
+        position_embs = self.position_embeddings(positions).expand_as(tensor)
+        tensor = tensor + position_embs
+
+        if self.n_segments >= 1:
+            if segments is None:
+                segments = torch.zeros_like(input)  # type: ignore
+            tensor = tensor + self.segment_embeddings(segments)
+
+        return tensor, mask
+
+    def forward_layers(
+        self, tensor: torch.Tensor, mask: torch.BoolTensor
+    ) -> torch.Tensor:
+        """
+        Apply transformer layers to input.
+
+        :param tensor:
+            embedded input
+        :param mask:
+            mask of input
+
+        :return tensor:
+            return embedding after applying transformer layers
+        """
+        if getattr(self.layers, 'is_model_parallel', False):
+            # factored out for readability. It is equivalent to the other
+            # condition
+            tensor = self._apply_model_parallel(tensor, mask)
+        else:
+            for i in range(self.n_layers):
+                tensor = self.layers[i](tensor, mask)
+
+        return tensor
+
+    def reduce_output(
+        self, tensor: torch.Tensor, mask: torch.BoolTensor
+    ) -> Tuple[torch.Tensor, Optional[torch.BoolTensor]]:
+        """
+        Reduce transformer output at end of forward pass.
+
+        :param tensor:
+            encoded input
+        :param mask:
+            mask for encoded input
+
+        :return (tensor, mask):
+            returns the reduced tensor, and mask if appropriate
+        """
+        tensor *= self.output_scaling
+        if self.reduction_type == 'first':
+            return tensor[:, 0, :], None
+        elif self.reduction_type == 'max':
+            return tensor.max(dim=1)[0], None
+        elif self.reduction_type == 'mean':
+            divisor = mask.float().sum(dim=1).unsqueeze(-1).clamp(min=1).type_as(tensor)
+            output = tensor.sum(dim=1) / divisor
+            return output, None
+        elif self.reduction_type is None or 'none' in self.reduction_type:
+            return tensor, mask
+        else:
+            raise ValueError(
+                "Can't handle --reduction-type {}".format(self.reduction_type)
+            )
+
+    def forward(  # type: ignore
+        self,
+        input: torch.LongTensor,
+        positions: Optional[torch.LongTensor] = None,
+        segments: Optional[torch.LongTensor] = None,
+    ) -> Union[torch.Tensor, Tuple[torch.Tensor, torch.BoolTensor]]:
+        """
+        Forward pass.
+
+        :param LongTensor[batch,seqlen] input:
+            The input IDs
+        :param LongTensor[batch,seqlen] positions:
+            Positions for input IDs
+        :param LongTensor[batch,seqlen] segments:
+            If provided, additionally adds ``segments`` as extra embedding features.
+        """
+        # embed input
+        tensor, mask = self.forward_embedding(input, positions, segments)
+
+        if self.variant == 'xlm' or self.variant == 'bart':
+            tensor = self.norm_embeddings(tensor)
+
+        # --dropout on the embeddings
+        tensor = self.dropout(tensor)
+
+        tensor *= mask.unsqueeze(-1).type_as(tensor)
+
+        # apply transformer layers
+        tensor = self.forward_layers(tensor, mask)
+
+        if self.variant == 'prelayernorm':
+            tensor = self.norm_embeddings(tensor)
+
+        # reduce output
+        tensor, out_mask = self.reduce_output(tensor, mask)
+        if out_mask is not None:
+            return tensor, out_mask
+        else:
+            return tensor
+
+    def _apply_model_parallel(self, tensor, mask):
+        """
+        Pipeline application of model parallelism.
+        """
+        chunks = PipelineHelper.split((tensor, mask))
+        work_items = PipelineHelper.schedule_work_items(self.layers, chunks)
+
+        for chunk_idx, layer_nos, next_device in work_items:
+            s_tensor, s_mask = chunks[chunk_idx]
+            for layer_no in layer_nos:
+                s_tensor = self.layers[layer_no](s_tensor, s_mask)
+            chunks[chunk_idx] = PipelineHelper.chunk_to((s_tensor, s_mask), next_device)
+
+        tensor_out, mask_out = PipelineHelper.join(chunks)
+        return tensor_out
+
+
+class TransformerEncoderLayer(nn.Module):
+    """
+    Implements a single Transformer encoder layer.
+    """
+
+    def __init__(
+        self,
+        n_heads,
+        embedding_size,
+        ffn_size,
+        attention_dropout=0.0,
+        relu_dropout=0.0,
+        dropout=0.0,
+        activation='relu',
+        variant=None,
+    ):
+        super().__init__()
+        self.dim = embedding_size
+        self.ffn_dim = ffn_size
+        self.activation = activation
+        self.variant = variant
+        self.attention = MultiHeadAttention(
+            n_heads, embedding_size, dropout=attention_dropout  # --attention-dropout
+        )
+        self.norm1 = torch.nn.LayerNorm(embedding_size, eps=LAYER_NORM_EPS)
+        self.ffn = TransformerFFN(
+            embedding_size,
+            ffn_size,
+            relu_dropout=relu_dropout,
+            activation=self.activation,
+        )
+        self.norm2 = torch.nn.LayerNorm(embedding_size, eps=LAYER_NORM_EPS)
+        self.dropout = nn.Dropout(p=dropout)
+
+    def forward(self, tensor, mask):
+        """
+        Forward pass.
+        """
+        residual = tensor
+        if self.variant == 'prelayernorm':
+            tensor = self.norm1(tensor)
+        attended_tensor = self.attention(tensor, mask=mask)[0]
+        tensor = residual + self.dropout(attended_tensor)
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            tensor = self.norm1(tensor)
+        residual = tensor
+        if self.variant == 'prelayernorm':
+            tensor = self.norm2(tensor)
+        tensor = residual + self.dropout(self.ffn(tensor))
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            tensor = self.norm2(tensor)
+        tensor *= mask.unsqueeze(-1).type_as(tensor)
+        return tensor
